# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A²ES - AI for Energy & Sustainability API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{equipmentId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get technical and ESG report for equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentReport"
        "404":
          description: No report found
  /reports/{equipmentId}/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate report with AI (technical + ESG + financing)
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentReport"
components:
  schemas:
...
    EquipmentReport:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        currentSituation:
          type: string
        problemsIdentified:
          type: string
        improvementOpportunities:
          type: string
        retrofitSolutions:
          type: string
        estimatedCostsBrl:
          type: number
          nullable: true
        consumptionReductionPct:
          type: number
          nullable: true
        co2ReductionKgMonth:
          type: number
          nullable: true
        financialGainsBrl:
          type: number
          nullable: true
        sustainableImpact:
          type: string
          nullable: true
        technicalJustification:
          type: string
          nullable: true
        financingReadiness:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - equipmentId
        - currentSituation
        - problemsIdentified
        - improvementOpportunities
        - retrofitSolutions
        - generatedAt